fix(plugins): preserve tabs during context refresh
